refactor(root): delegate registry registration to global-db
