+++ crates/tracedecay/src/tracedecay.rs
@@ -27,7 +27,6 @@ mod source_edit_runtime;
 
 pub use diagnostics::{BranchDiagnostics, TrackedBranchDiagnostic};
 pub use lifecycle::MovedStoreAdoption;
-pub(crate) use lifecycle::git_remote_url;
 
 /// Central orchestrator that coordinates all subsystems of the code graph.
 ///
